backend2/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
openai==1.3.0
werkzeug==2.3.7
flask_cors
gtts
transformers
torch
diffusers 
librosa
pydub